requirements.txt
appdirs==1.4.4
asgiref==3.3.4
black==21.5b1
certifi==2021.5.30
chardet==4.0.0
click==8.0.0
coreapi==2.3.3
coreschema==0.0.4
Django==3.2.3
django-rest-framework==0.1.0
django-taggit==1.4.0
djangorestframework==3.12.4
drf-yasg2==1.19.4
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.1
MarkupSafe==2.0.1
mypy-extensions==0.4.3
packaging==20.9
pathspec==0.8.1
pyparsing==2.4.7
pytz==2021.1
regex==2021.4.4
requests==2.25.1
ruamel.yaml==0.17.7
ruamel.yaml.clib==0.2.2
six==1.16.0
sqlparse==0.4.1
toml==0.10.2
uritemplate==3.0.1
urllib3==1.26.5
